adding OOp notes and comments to mortgage tutorial

+++ MortgageTutorial/src/com/mortgage/tutorial/Console.java
@@ -2,11 +2,13 @@ package com.mortgage.tutorial;
 
 import java.util.Scanner;
 
-//create a class to read and display the input being entered in the terminal
+//create a class to read and store the input being entered in the terminal
 public class Console {
 
-	// create a scanner and declare it private static as this is the only one and
-	// the only one used in this project
+	/**
+	 * create a scanner and declare it private static as this is the only one and
+	 * the only one used in this project
+	 */
 	private static Scanner scanner = new Scanner(System.in);
 
 	public static double readNumber(String prompt, double min, double max) {
@@ -14,7 +16,7 @@ public class Console {
 		double value;
 		while (true) {
 			System.out.println(prompt);
-			value = scanner.nextInt();
+			value = scanner.nextDouble();
 
 			if (value >= min && value <= max) {
 				break;
